requirements.txt
pygame>=2.0.0
requests>=2.25.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=0.4.0
google-auth>=1.0.0
oauth2client>=4.1.0
psutil>=5.8.0
python-dotenv>=1.0.0 